Task #7332. Use Money from waCommon
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@11481 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ frameworks/common-ui/src/main/java/ish/oncourse/ui/components/CourseClassPrice.java
@@ -1,6 +1,6 @@
 package ish.oncourse.ui.components;
 
-import ish.oncourse.math.Money;
+import ish.math.Money;
 import ish.oncourse.model.CourseClass;
 import ish.oncourse.model.Discount;
 import ish.oncourse.model.DiscountCourseClass;
@@ -62,7 +62,7 @@ public class CourseClassPrice {
 		if (discountedFee == null) {
 			discountedFee = Money.valueOf(courseClass.getFeeExGst().subtract(
 					getDiscountValue())
-					.multiply(courseClass.getTaxMultiplier()));
+					.multiply(courseClass.getTaxMultiplier())).toBigDecimal();
 
 			if (discountedFee == null) {
 				discountedFee = Money.ZERO.toBigDecimal();
@@ -131,7 +131,7 @@ public class CourseClassPrice {
 				.subtract(
 						Discount.discountValueForCourseClass(discountsList,
 								courseClass)).multiply(
-						courseClass.getTaxMultiplier()));
+						courseClass.getTaxMultiplier())).toBigDecimal();
 	}
 
 }
